x86/msr: Add testcases for MSR_IA32_PRED_CMD and its IBPB command
Add test coverage to verify MSR_IA32_PRED_CMD is write-only, that it can
be written with '0' (nop command) and '1' (IBPB command) when IBPB is
supported by the CPU (SPEC_CTRL on Intel, IBPB on AMD).  Do not check
that writing any other bit (1-63) triggers a #GP due to the bits/commands
being reserved, because some versions of the microcode do not do that.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20230328050231.3008531-3-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>