s390x/snippets: Fix compilation with Clang 15
Clang complains:

 s390x/snippets/c/cstart.S:22:13: error: invalid operand for instruction
  lghi %r15, stackptr
             ^
Let's load the address with "larl" instead, like we already do
it in s390x/cstart64.S. For this we should also switch to 64-bit
mode first, then we also don't have to clear r15 right in front
of this anymore.

Changing the code here triggered another problem: initial_cr0
must be aligned on a double-word boundary, otherwise the lctlg
instruction will fail with an specification exception. This was
just working by accident so far - add an ".align 8" now to avoid
the problem.

Message-Id: <20230404101434.172721-1-thuth@redhat.com>
Reviewed-by: Nico Boehr <nrb@linux.ibm.com>
Reviewed-by: Janosch Frank <frankja@linux.ibm.com>
Reviewed-by: Claudio Imbrenda <imbrenda@linux.ibm.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>