MAINTAINERS: Add a catch-all entry for the kvm mailing list
The scripts/get_maintainer.pl currently fails to suggest sending
patches to kvm@vger.kernel.org if a patch only touches files that
are not part of any target specific code (e.g. files in the script
folder). All patches should be CC:-ed to the kvm list, so we should
have an entry here that covers all files.

Message-ID: <20230404125103.200027-1-thuth@redhat.com>
Acked-by: Janosch Frank <frankja@linux.ibm.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>