x86: Drop VPID invl tests from nested reduced MAXPHYADDR access testcase
Remove the variants of the nested VMX reduced MAXPHYADDR access test that
require a nested VM-Exit in order to invalidate the to-be-tested
translation, as the combination is quite uninteresting, and those tests
have a painfully long runtime (upwards of 10+ minutes).  Alternatively,
the tests could be split out a la commit ca785dae ("vmx: separate VPID
tests"), but the probability of finding a KVM bug that only affects
VPID-based invladations in combination with a smaller MAXPHYADDR is
extremely low.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Message-Id: <20230404165341.163500-8-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>