lib/efi: Add support for reading an FDT
This change adds support for reading environment variables. To do so it
introduces a new GUID: 97ef3e03-7329-4a6a-b9ba-6c1fdcc5f823 that the
user needs to provide when setting environment variables. For example,
to set the path to the fdt a user can execute:

In addition, this change add support for reading the fdt into memory
and providing a pointer to the test through efi_bootinfo_t.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
[Ensure DTB pathname is nul-terminated.]
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>