arm64: Use the provided fdt when booting through EFI
This patch uses the fdt pointer provided through efi_bootinfo_t for
tests that run as EFI apps. As in Linux, we give priority to the fdt.
First, we check if the pointer to the fdt is set and only if it isn't
we fallback to using the ACPI.

In addition, this patches changes the efi run script to generate and
use a fdt unless the user has set the enviroment variable EFI_USE_ACPI
to 'y'.

As a result:

$> ./arm/efi/run ./arm/selftest.efi -append "setup smp=2 mem=256" -smp 2 -m 256

will use an fdt, where as

$> EFI_USE_ACPI=y ./arm/efi/run ./arm/selftest.efi -append "setup smp=2 mem=256" -smp 2 -m 256

will run relying on ACPI.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>