runtime: don't run pv-host tests when gen-se-header is unavailable
When the gen-se-header tool is not given as an argument to configure,
all tests which act as a PV host will not be built by the makefiles.

run_tests.sh will fail when a test binary is missing. This means
when we add the pv-host tests to unittest.cfg we will have FAILs when
gen-se-header is missing.

Since it is desirable to have the tests in unittest.cfg, add a new group
pv-host which designates tests that act as a PV host. These will only
run if the gen-se-header tool is available.

The pv-host group is currently not used, but will be with Janoschs
series "s390x: Add PV SIE intercepts and ipl tests" here:
https://lore.kernel.org/all/20230502115931.86280-1-frankja@linux.ibm.com/

Reviewed-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Nico Boehr <nrb@linux.ibm.com>