runtime: Allow to specify properties for accelerator
There are extra properties for accelerators to enable the specific
features. For example, the dirty ring for KVM accelerator can be
enabled by "-accel kvm,dirty-ring-size=65536". Unfortuntely, the
extra properties for the accelerators aren't supported. It makes
it's impossible to test the combination of KVM and dirty ring
as the following error message indicates.

  # cd /home/gavin/sandbox/kvm-unit-tests/tests
  # QEMU=/home/gavin/sandbox/qemu.main/build/qemu-system-aarch64 \
    ACCEL=kvm,dirty-ring-size=65536 ./its-migration
     :
  BUILD_HEAD=2fffb37e
  timeout -k 1s --foreground 90s                                 \
  /home/gavin/sandbox/qemu.main/build/qemu-system-aarch64        \
  -nodefaults -machine virt -accel kvm,dirty-ring-size=65536     \
  -cpu cortex-a57 -device virtio-serial-device                   \
  -device virtconsole,chardev=ctd -chardev testdev,id=ctd        \
  -device pci-testdev -display none -serial stdio                \
  -kernel _NO_FILE_4Uhere_ -smp 160 -machine gic-version=3       \
  -append its-pending-migration # -initrd /tmp/tmp.gfDLa1EtWk
     :
  qemu-system-aarch64: kvm_init_vcpu: kvm_arch_init_vcpu failed (0):
  Invalid argument

Allow to specify extra properties for accelerators. With this, the
"its-migration" can be tested for the combination of KVM and dirty
ring. Rename get_qemu_accelerator() to set_qemu_accelerator() since
no values are returned by printing at return.

Signed-off-by: Gavin Shan <gshan@redhat.com>
Tested-by: Nico Boehr <nrb@linux.ibm.com>
Acked-by: Nico Boehr <nrb@linux.ibm.com>
Reviewed-by: Andrew Jones <andrew.jones@linux.dev>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>