migration: Fix test harness hang if source does not reach migration point
After starting the test, the harness waits polling for "migrate" in the
output. If the test does not print for some reason, the harness hangs.

Test that the pid is still alive while polling to fix this hang.

While here, wait for the full string "Now migrate the VM", which I think
makes it more obvious to read and could avoid an unfortunate collision
with some debugging output in a test case.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Message-ID: <20230725033937.277156-3-npiggin@gmail.com>
Reviewed-by: Nico Boehr <nrb@linux.ibm.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>