nVMX: Drop testcase that falsely claims to verify vmcs.HOST_RIP[63:32]
Excise the completely bogus testcase in test_host_addr_size() which
purports to verify that setting vmcs.HOST_RIP[63:32] to non-zero values is
allowed for 64-bit hosts.  The testcase is mindbogglingly broken: setting
arbitrary, single bits above bit 46 creates a noncanonical address, and
setting arbitrary bits below bit 47 would send the test into the weeds as
a "successful" VMLAUNCH generates a VM-Exit, i.e. would load the garbage
RIP and immediately encounter a #PF.

The only reason the passes is because it does absolutely nothing useful:
vmlaunch() unconditionally writes HOST_RIP before VMLAUNCH, because not
jumping to a random RIP on a VM-Exit is mildly important.

Outright drop the testcase, trying to salvage anything from it would be a
waste of time as simply running any 64-bit guest will generate a huge
variety of RIPs with non-zero values in bits 63:32.

Link: https://lore.kernel.org/r/20230901225004.3604702-7-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>