arm: PMU: Fix test condition
The test description says: "check overflow happened on #0 only",
but the test condition actually tests: "check overflow happened
AT LEAST of #0, maybe of some other counters as well."
This basically disregards any overflow of other counters, but it
should ensure that there are none.

With the updated test condition the test will fail if there was an
overflow of counter #1, too.

Fixes: 4ce2a8045624 ("arm: pmu: Basic event counter Tests")
Signed-off-by: Matthias Rosenfelder <matthias.rosenfelder@nio.io>
Reviewed-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>