(arm|powerpc|s390x): Makefile: Fix .aux.o generation
Using all prerequisites for the source file results in the build
dying on the second time around with:

gcc: fatal error: cannot specify ‘-o’ with ‘-c’, ‘-S’ or ‘-E’ with multiple files

This is due to auxinfo.h becoming a prerequisite after the first
build recorded the dependency.

Use the first prerequisite for this recipe.

Fixes: f2372f2d49135 ("(arm|powerpc|s390x): Makefile: add `%.aux.o` target")
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Message-ID: <20240202065740.68643-2-npiggin@gmail.com>
[thuth: Fix second occurance in arm/Makefile.common]
Signed-off-by: Thomas Huth <thuth@redhat.com>