lib/arm/io: Fix calling getchar() multiple times
getchar() can currently only be called once on arm since the implementation
is a little bit too  naïve: After the first character has arrived, the
data register never gets set to zero again. To properly check whether a
byte is available, we need to check the "RX fifo empty" on the pl011 UART
or the "RX data ready" bit on the ns16550a UART instead.

With this proper check in place, we can finally also get rid of the
ugly assert(count < 16) statement here.

Message-ID: <20240216140210.70280-1-thuth@redhat.com>
Acked-by: Andrew Jones <andrew.jones@linux.dev>
Tested-by: Alexandru Elisei <alexandru.elisei@arm.com>
Tested-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>